+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 
 - Permit zero submission cost and submissions (also to ease testing of submission creation).
 - Fixed #45: Sorting and filtering submissions.
+- Fixed #28: Moved invitations to scholar page.
 
 ### Fixed